Perubahan UI/UX

+++ app/Http/Controllers/admin/UsersController.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\admin;
 
+use App\Http\Controllers\Controller;
+use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -27,9 +29,9 @@ class UsersController extends Controller
             }
         }
 
-        $users = $query->paginate(2)->appends($request->only('search', 'role'));
+        $users = $query->paginate(5)->appends($request->only('search', 'role'));
 
-        return view('admin.kelola_user', compact('users'));
+        return view('admin.kelola_user.index', compact('users'));
     }
 
 
@@ -38,23 +40,29 @@ class UsersController extends Controller
      */
     public function create()
     {
-        return view('admin.__create');
+        return view('admin.kelola_user.__create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(UserRequest $request)
     {
+        $data = $request->validated();
 
         $data['name'] = $request->name;
         $data['email'] = $request->email;
         $data['password'] = Hash::make($request->password);
-        $data['role'] = $request->role;
+
+        if ($request->role) {
+            $data['role'] = $request->role;
+        } else {
+            $data['role'] = 'pelanggan';
+        }
 
         User::create($data);
 
-        return redirect()->route('user.index');
+        return redirect()->route('adminadminuser.index')->with('success', 'User berhasil ditambahkan!');
     }
 
     /**
@@ -71,24 +79,32 @@ class UsersController extends Controller
     public function edit(string $id)
     {
         $user = User::find($id);
-        return view('admin.__edit', compact('user'));
+        if (!$user) {
+            return redirect()->route('adminuser.index')->with('error', 'User tidak ditemukan!');
+        }
+        return view('admin.kelola_user.__edit', compact('user'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(UserRequest $request, string $id)
     {
-        $data = User::find($id);
+        $find = User::find($id);
+
+        $data = $request->validated();
 
         $data['name'] = $request->name;
         $data['email'] = $request->email;
-        $data['password'] = Hash::make($request->password);
+        if (!empty($data['password'])) {
+            $data['password'] = Hash::make($data['password']);
+        } else {
+            unset($data['password']);
+        }
         $data['role'] = $request->role;
+        $find->update($data);
 
-        $data->update();
-
-        return redirect()->route('adminadmin.index');
+        return redirect()->route('adminadminuser.index')->with('success', 'User berhasil diperbarui!');
     }
 
     /**
@@ -98,11 +114,12 @@ class UsersController extends Controller
     {
         $data = User::find($id);
 
-        try {
-            $data->delete();
-            return redirect()->route('user.index')->with('success', 'User berhasil dihapus!');
-        } catch (\Exception $e) {
-            return redirect()->route('user.index')->with('error', 'Terjadi kesalahan saat menghapus user: ' . $e->getMessage());
+        if (!$data) {
+            return redirect()->route('adminadminuser.index')->with('error', 'User tidak ditemukan!');
         }
+
+        $data->delete();
+
+        return redirect()->route('adminadminuser.index')->with('success', 'User berhasil dihapus!');
     }
 }
